+++ source/evael/renderer/vk/vk_buffer.d
@@ -39,7 +39,7 @@ class VulkanBuffer(BufferType type) : Buffer!type
 	 *		 data : pointer to the new data that will be copied into the data store
 	 */
 	@nogc
-	public override void update(in long offset, in ptrdiff_t size, in void* data) const nothrow
+	public override void update(in ptrdiff_t offset, in ptrdiff_t size, in void* data) const nothrow
 	{
 		assert(offset + size <= this.m_size, "Updating buffer with invalid offset/size.");
 	}
